+ first version of caret/encoding history
