+++ src/Edit.h
@@ -112,77 +112,9 @@ void  EditMarkAll(HWND,int,BOOL,BOOL);
 void  EditSetAccelWordNav(HWND,BOOL);
 void  CompleteWord(HWND,BOOL);
 
-extern int g_DOSEncoding;
-
-#define NCP_DEFAULT            1
-#define NCP_UTF8               2
-#define NCP_UTF8_SIGN          4
-#define NCP_UNICODE            8
-#define NCP_UNICODE_REVERSE   16
-#define NCP_UNICODE_BOM       32
-#define NCP_8BIT              64
-#define NCP_ANSI             128
-#define NCP_OEM              256
-#define NCP_MBCS             512
-#define NCP_INTERNAL          (NCP_DEFAULT|NCP_UTF8|NCP_UTF8_SIGN|NCP_UNICODE|NCP_UNICODE_REVERSE|NCP_UNICODE_BOM|NCP_ANSI|NCP_OEM|NCP_MBCS)
-#define NCP_RECODE           1024
-
-#define CPI_GET               -2
-#define CPI_NONE              -1
-#define CPI_ANSI_DEFAULT       0
-#define CPI_OEM                1
-#define CPI_UNICODEBOM         2
-#define CPI_UNICODEBEBOM       3
-#define CPI_UNICODE            4
-#define CPI_UNICODEBE          5
-#define CPI_UTF8               6
-#define CPI_UTF8SIGN           7
-#define CPI_UTF7               8
-
-#define IDS_ENCODINGNAME0  61000
-#define IDS_EOLMODENAME0   62000
-
-typedef struct _np2encoding {
-  UINT    uFlags;
-  UINT    uCodePage;
-  char*   pszParseNames;
-  int     idsName;
-  WCHAR   wchLabel[64];
-} NP2ENCODING;
-
-int  Encoding_Current(int);    // getter/setter
-int  Encoding_Source(int);     // getter/setter
-int  Encoding_SrcWeak(int);    // getter/setter
-BOOL Encoding_HasChanged(int); // query/setter
-
-void Encoding_InitDefaults();
-int  Encoding_MapIniSetting(BOOL,int);
-void Encoding_GetLabel(int);
-int  Encoding_MatchW(LPCWSTR);
-int  Encoding_MatchA(char*);
-BOOL Encoding_IsValid(int);
-int  Encoding_GetByCodePage(UINT);
-void Encoding_AddToListView(HWND,int,BOOL);
-BOOL Encoding_GetFromListView(HWND,int *);
-void Encoding_AddToComboboxEx(HWND,int,BOOL);
-BOOL Encoding_GetFromComboboxEx(HWND,int *);
-BOOL Encoding_IsDefault(int);
-BOOL Encoding_IsANSI(int);
-BOOL Encoding_IsOEM(int);
-
-UINT Encoding_SciGetCodePage(HWND);
-int  Encoding_SciMappedCodePage(int);
-void Encoding_SciSetCodePage(HWND,int);
-
-
-BOOL IsUnicode(const char*,int,LPBOOL,LPBOOL);
-BOOL IsUTF8(const char*,int);
-BOOL IsUTF7(const char*,int);
-
 
 //void SciInitThemes(HWND);
 //LRESULT CALLBACK SciThemedWndProc(HWND,UINT,WPARAM,LPARAM);
-
 
 #define FV_TABWIDTH        1
 #define FV_INDENTWIDTH     2
